[irdl] Move TypeConstraint in a separate file
